str updates

+++ Metrics.py
@@ -27,8 +27,8 @@ def triplet(Y_Target, Y_Predict, multi_dimension = False, output_format = list):
     Accepts appended results in each iterations. 
     Put multi_dimension = Ture
     """
-    Y_Target = np.asarray(Y_Target,order = 'C')
-    Y_Predict = np.asarray(Y_Predict,order = 'C')
+    Y_Target = np.asarray(Y_Target,order = 'C',dtype=float)
+    Y_Predict = np.asarray(Y_Predict,order = 'C',dtype=float)
     if multi_dimension:
         Y_Target = Y_Target.flatten()
         Y_Predict = Y_Predict.flatten()
